minor refactor to fix stubs
