+++ server/test/unit/services/cssr-records/cssr-record.spec.js
@@ -2,6 +2,7 @@ const expect = require('chai').expect;
 const sinon = require('sinon');
 
 const cssrRecord = require('../../../../services/cssr-records/cssr-record');
+const cssrRecordData = require('../../../../services/cssr-records/cssrRecordData');
 
 const la = {
   theAuthority: {
@@ -14,14 +15,17 @@ const la = {
 describe('/server/services/cssr-records/cssr-record', async () => {
   describe('GetCssrRecordFromPostcode', async () => {
     it('should return a cssr record when a matching postcode with a corresponding cssr entry is found', async () => {
-      sinon.stub(cssrRecord, 'getCssrRecordCompleteMatch').callsFake(async (args) => {
-        return { postcode: args.where.postcode, ...la };
+      const stubCompleteMatch = sinon.stub(cssrRecordData, 'getCssrRecordCompleteMatch').callsFake(async () => {
+        return la;
       });
-      // let spy = sinon.spy(cssrRecord);
+      const stubSimilarMatch = sinon.stub(cssrRecordData, 'getCssrRecordWithLikePostcode').callsFake(async () => {
+        console.log('This should not be hit');
+      });
 
       const localAuth = await cssrRecord.GetCssrRecordFromPostcode('HD1 1DA');
-      expect(localAuth).to.deep.equal(la.theAuthority.name);
-      expect(sinon.getCssrRecordCompleteMatch.calledOnce);
+      expect(localAuth).to.deep.equal(la);
+      expect(stubCompleteMatch.calledOnce);
+      expect(stubSimilarMatch.notCalled);
     });
   });
 });
